+++ src/test/java/com/braintreegateway/unittest/TransactionTest.java
@@ -364,4 +364,33 @@ public class TransactionTest {
 
 		assertTrue(transaction.isforeignRetailer());
 	}
+
+    @Test
+	public void parseInternationalPhoneTransaction() {
+		String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" +
+				"<transaction>\n" +
+				"  <type>sale</type>\n" +
+				"  <customer></customer>\n" +
+				"  <billing>\n" +
+                "    <international-phone>\n" +
+                "      <country-code>1</country-code>\n" +
+                "      <national-number>3121234567</national-number>\n" +
+                "    </international-phone>\n" +
+                "  </billing>\n" +
+				"  <shipping>\n" +
+                "    <international-phone>\n" +
+                "      <country-code>2</country-code>\n" +
+                "      <national-number>3121234568</national-number>\n" +
+                "    </international-phone>\n" +
+                "  </shipping>\n" +
+				"</transaction>\n";
+
+		SimpleNodeWrapper transactionNode = SimpleNodeWrapper.parse(xml);
+		Transaction transaction = new Transaction(transactionNode);
+
+		assertEquals("1", transaction.getBillingAddress().getInternationalPhone().getCountryCode());
+		assertEquals("3121234567", transaction.getBillingAddress().getInternationalPhone().getNationalNumber());
+		assertEquals("2", transaction.getShippingAddress().getInternationalPhone().getCountryCode());
+		assertEquals("3121234568", transaction.getShippingAddress().getInternationalPhone().getNationalNumber());
+	}
 }